+++ src/widgets/VKID/index.ts
