+++ src/widgets/VKID/model/types.ts
