+++ src/widgets/VKID/ui/VKIDPopup.hbs
