+++ src/widgets/VKID/ui/VKIDPopup.scss
